Refactor: separate api and client projects
