+++ api/src/main/scala/hmda/geo/http/Service.scala
@@ -1,8 +1,8 @@
-package hmda.geo.api
+package hmda.geo.http
 
 import akka.http.scaladsl.server.Directives._
-import hmda.geo.api.routes._
-import hmda.geo.api.routes.census._
+import hmda.geo.http.routes._
+import hmda.geo.http.routes.census._
 
 trait Service extends StatusRoute with TractRoute {
   val routes = statusRoute ~ tractRoute
