+++ api/src/main/scala/hmda/geo/http/routes/StatusRoute.scala
@@ -1,4 +1,4 @@
-package hmda.geo.api.routes
+package hmda.geo.http.routes
 
 import java.net.InetAddress
 import java.time.Instant
