+++ api/src/main/scala/hmda/geo/model/GeometryTable.scala
