+++ api/src/test/scala/hmda/geo/protocol/StatusJsonProtocolSpec.scala
